Tell boost to not request linking to tis libraries...
MSVC allows adding comments to object files via "#pragma comment(...)".
This feature also allows to add comments to the object files that tell
the linker which libraries to link in additionally, whenever such an
object file is linked into a library.
Boost uses this in its headers. Any code using boost headers that require
boost libraries therefore will fail when linked if the library cannot be
found. STP's build system already takes care of linking boost, so this is
not needed in our case. If this automatic linking fails for some reason
it is not readily apparent why this failed.

This patch tells boost not to insert these pragmas into the object files.