<msvc/stdint.h> -> "msvc/stdint.h" ...
in various headers in the cryptominisat2 directory. msvc/ is a folder local
to cryptominisat. This implementation of stdint.h is necessary because
MSVC doesn't provide this header at all.

Note that there is another implementaion of stdin.h in STP's windows/
directory. I have no idea what happens if these differ.