make inline work for visual studio...
this is odd: On the one hand, inline is not a keyword in visual studio so
it doesn't support it. On the other hand it doesn't allow you to redefine
it because it is a keyword. Fortunately we can just tell visual studio to
allow keyword macros and it shuts up about this.